+++ examples/react-apollo/src/components/OauthLinks.tsx
@@ -1,0 +1,23 @@
+import React from 'react'
+import { FaFacebook, FaGithub, FaGoogle } from 'react-icons/fa'
+
+import { useProviderLink } from '@nhost/react'
+
+import AuthLink from './AuthLink'
+
+export default function OauthLinks() {
+  const { github, google, facebook } = useProviderLink()
+  return (
+    <>
+      <AuthLink icon={<FaGithub />} link={github} color="#333">
+        Continue with GitHub
+      </AuthLink>
+      <AuthLink icon={<FaGoogle />} link={google} color="#de5246">
+        Continue with Google
+      </AuthLink>
+      <AuthLink icon={<FaFacebook />} link={facebook} color="#3b5998">
+        Continue with Facebook
+      </AuthLink>
+    </>
+  )
+}