+++ packages/core/src/machines/index.typegen.ts
@@ -5,19 +5,9 @@ export interface Typegen0 {
   eventsCausingActions: {
     saveSession:
       | 'SESSION_UPDATE'
+      | 'done.invoke.importRefreshToken'
       | 'done.invoke.authenticatePasswordlessSmsOtp'
       | 'done.invoke.authenticateUserWithPassword'
-      | 'done.invoke.signInToken'
-      | 'done.invoke.authenticateAnonymously'
-      | 'done.invoke.signInMfaTotp'
-      | 'done.invoke.signUpUser'
-      | 'done.invoke.refreshToken'
-      | 'done.invoke.authenticateWithToken'
-    persist:
-      | 'SESSION_UPDATE'
-      | 'done.invoke.authenticatePasswordlessSmsOtp'
-      | 'done.invoke.authenticateUserWithPassword'
-      | 'done.invoke.signInToken'
       | 'done.invoke.authenticateAnonymously'
       | 'done.invoke.signInMfaTotp'
       | 'done.invoke.signUpUser'
@@ -26,31 +16,36 @@ export interface Typegen0 {
     resetTimer: 'SESSION_UPDATE' | 'done.invoke.refreshToken' | ''
     reportTokenChanged:
       | 'SESSION_UPDATE'
+      | 'done.invoke.importRefreshToken'
       | 'done.invoke.authenticatePasswordlessSmsOtp'
       | 'done.invoke.authenticateUserWithPassword'
-      | 'done.invoke.signInToken'
       | 'done.invoke.authenticateAnonymously'
       | 'done.invoke.signInMfaTotp'
       | 'done.invoke.signUpUser'
       | 'done.invoke.refreshToken'
       | 'done.invoke.authenticateWithToken'
-    saveRefreshToken: 'done.invoke.importRefreshToken'
     saveAuthenticationError:
       | 'error.platform.importRefreshToken'
       | 'error.platform.authenticatePasswordlessEmail'
       | 'error.platform.authenticatePasswordlessSms'
       | 'error.platform.authenticatePasswordlessSmsOtp'
       | 'error.platform.authenticateUserWithPassword'
-      | 'error.platform.signInToken'
       | 'error.platform.authenticateAnonymously'
       | 'error.platform.signInMfaTotp'
+    reportAwaitEmailVerification:
+      | 'done.invoke.authenticatePasswordlessEmail'
+      | 'error.platform.authenticateUserWithPassword'
+      | 'done.invoke.signUpUser'
+      | 'error.platform.signUpUser'
     saveMfaTicket: 'done.invoke.authenticateUserWithPassword'
-    broadcastToken: 'done.invoke.signInToken'
     saveRegisrationError: 'error.platform.signUpUser'
     saveRefreshAttempt: 'error.platform.refreshToken'
     reportSignedOut:
       | 'error.platform.importRefreshToken'
-      | ''
+      | 'done.invoke.authenticatePasswordlessEmail'
+      | 'error.platform.authenticateUserWithPassword'
+      | 'done.invoke.signUpUser'
+      | 'error.platform.signUpUser'
       | 'error.platform.authenticateWithToken'
     resetAuthenticationError: 'xstate.init'
     destroyRefreshToken: 'xstate.init'
@@ -58,26 +53,41 @@ export interface Typegen0 {
     resetSignUpError: 'SIGNUP_EMAIL_PASSWORD'
     reportSignedIn:
       | 'SESSION_UPDATE'
+      | 'done.invoke.importRefreshToken'
       | ''
       | 'done.invoke.authenticatePasswordlessSmsOtp'
       | 'done.invoke.authenticateUserWithPassword'
-      | 'done.invoke.signInToken'
       | 'done.invoke.authenticateAnonymously'
       | 'done.invoke.signInMfaTotp'
       | 'done.invoke.signUpUser'
       | 'done.invoke.authenticateWithToken'
     cleanUrl:
       | 'SESSION_UPDATE'
+      | 'done.invoke.importRefreshToken'
       | ''
       | 'done.invoke.authenticatePasswordlessSmsOtp'
       | 'done.invoke.authenticateUserWithPassword'
-      | 'done.invoke.signInToken'
+      | 'done.invoke.authenticateAnonymously'
+      | 'done.invoke.signInMfaTotp'
+      | 'done.invoke.signUpUser'
+      | 'done.invoke.authenticateWithToken'
+    broadcastToken:
+      | 'SESSION_UPDATE'
+      | 'done.invoke.importRefreshToken'
+      | ''
+      | 'done.invoke.authenticatePasswordlessSmsOtp'
+      | 'done.invoke.authenticateUserWithPassword'
       | 'done.invoke.authenticateAnonymously'
       | 'done.invoke.signInMfaTotp'
       | 'done.invoke.signUpUser'
       | 'done.invoke.authenticateWithToken'
   }
   internalEvents: {
+    'done.invoke.importRefreshToken': {
+      type: 'done.invoke.importRefreshToken'
+      data: unknown
+      __tip: 'See the XState TS docs to learn how to strongly type this.'
+    }
     'done.invoke.authenticatePasswordlessSmsOtp': {
       type: 'done.invoke.authenticatePasswordlessSmsOtp'
       data: unknown
@@ -85,11 +95,6 @@ export interface Typegen0 {
     }
     'done.invoke.authenticateUserWithPassword': {
       type: 'done.invoke.authenticateUserWithPassword'
-      data: unknown
-      __tip: 'See the XState TS docs to learn how to strongly type this.'
-    }
-    'done.invoke.signInToken': {
-      type: 'done.invoke.signInToken'
       data: unknown
       __tip: 'See the XState TS docs to learn how to strongly type this.'
     }
@@ -119,11 +124,6 @@ export interface Typegen0 {
       __tip: 'See the XState TS docs to learn how to strongly type this.'
     }
     '': { type: '' }
-    'done.invoke.importRefreshToken': {
-      type: 'done.invoke.importRefreshToken'
-      data: unknown
-      __tip: 'See the XState TS docs to learn how to strongly type this.'
-    }
     'error.platform.importRefreshToken': {
       type: 'error.platform.importRefreshToken'
       data: unknown
@@ -144,12 +144,16 @@ export interface Typegen0 {
       type: 'error.platform.authenticateUserWithPassword'
       data: unknown
     }
-    'error.platform.signInToken': { type: 'error.platform.signInToken'; data: unknown }
     'error.platform.authenticateAnonymously': {
       type: 'error.platform.authenticateAnonymously'
       data: unknown
     }
     'error.platform.signInMfaTotp': { type: 'error.platform.signInMfaTotp'; data: unknown }
+    'done.invoke.authenticatePasswordlessEmail': {
+      type: 'done.invoke.authenticatePasswordlessEmail'
+      data: unknown
+      __tip: 'See the XState TS docs to learn how to strongly type this.'
+    }
     'error.platform.signUpUser': { type: 'error.platform.signUpUser'; data: unknown }
     'error.platform.refreshToken': { type: 'error.platform.refreshToken'; data: unknown }
     'error.platform.authenticateWithToken': {
@@ -166,11 +170,6 @@ export interface Typegen0 {
       __tip: 'See the XState TS docs to learn how to strongly type this.'
     }
     'error.platform.signingOut': { type: 'error.platform.signingOut'; data: unknown }
-    'done.invoke.authenticatePasswordlessEmail': {
-      type: 'done.invoke.authenticatePasswordlessEmail'
-      data: unknown
-      __tip: 'See the XState TS docs to learn how to strongly type this.'
-    }
     'done.invoke.authenticatePasswordlessSms': {
       type: 'done.invoke.authenticatePasswordlessSms'
       data: unknown
@@ -184,13 +183,10 @@ export interface Typegen0 {
     signInPasswordlessSms: 'done.invoke.authenticatePasswordlessSms'
     signInPasswordlessSmsOtp: 'done.invoke.authenticatePasswordlessSmsOtp'
     signInPassword: 'done.invoke.authenticateUserWithPassword'
-    refreshToken:
-      | 'done.invoke.signInToken'
-      | 'done.invoke.refreshToken'
-      | 'done.invoke.authenticateWithToken'
     signInAnonymous: 'done.invoke.authenticateAnonymously'
     signInMfaTotp: 'done.invoke.signInMfaTotp'
     signUpUser: 'done.invoke.signUpUser'
+    refreshToken: 'done.invoke.refreshToken' | 'done.invoke.authenticateWithToken'
   }
   missingImplementations: {
     actions: never
@@ -200,7 +196,6 @@ export interface Typegen0 {
   }
   eventsCausingServices: {
     importRefreshToken: 'xstate.init'
-    refreshToken: '' | 'TRY_TOKEN'
     signInPassword: 'SIGNIN_PASSWORD'
     signInPasswordlessEmail: 'SIGNIN_PASSWORDLESS_EMAIL'
     signInPasswordlessSms: 'SIGNIN_PASSWORDLESS_SMS'
@@ -209,28 +204,26 @@ export interface Typegen0 {
     signInAnonymous: 'SIGNIN_ANONYMOUS'
     signInMfaTotp: 'SIGNIN_MFA_TOTP'
     signout: 'SIGNOUT'
+    refreshToken: '' | 'TRY_TOKEN'
   }
   eventsCausingGuards: {
     hasSession: 'SESSION_UPDATE' | 'done.invoke.signUpUser'
     isSignedIn: '' | 'error.platform.authenticateWithToken'
-    hasRefreshTokenWithoutSession: ''
-    hasAuthenticationError: ''
     hasMfaTicket: 'done.invoke.authenticateUserWithPassword'
     unverified: 'error.platform.authenticateUserWithPassword' | 'error.platform.signUpUser'
     noToken: ''
     isAutoRefreshDisabled: ''
     hasRefreshToken: ''
     refreshTimerShouldRefresh: ''
+    needsVerification: 'SIGNED_IN'
   }
   eventsCausingDelays: {}
   matchesStates:
     | 'authentication'
-    | 'authentication.importingRefreshToken'
     | 'authentication.starting'
     | 'authentication.signedOut'
     | 'authentication.signedOut.noErrors'
     | 'authentication.signedOut.success'
-    | 'authentication.signedOut.needsEmailVerification'
     | 'authentication.signedOut.needsSmsOtp'
     | 'authentication.signedOut.needsMfa'
     | 'authentication.signedOut.failed'
@@ -245,7 +238,6 @@ export interface Typegen0 {
     | 'authentication.authenticating.passwordlessSms'
     | 'authentication.authenticating.passwordlessSmsOtp'
     | 'authentication.authenticating.password'
-    | 'authentication.authenticating.token'
     | 'authentication.authenticating.anonymous'
     | 'authentication.authenticating.mfa'
     | 'authentication.authenticating.mfa.totp'
@@ -266,9 +258,11 @@ export interface Typegen0 {
     | 'token.idle.noErrors'
     | 'token.idle.error'
     | 'token.running'
+    | 'email'
+    | 'email.awaitingVerification'
+    | 'email.valid'
     | {
         authentication?:
-          | 'importingRefreshToken'
           | 'starting'
           | 'signedOut'
           | 'authenticating'
@@ -278,7 +272,6 @@ export interface Typegen0 {
               signedOut?:
                 | 'noErrors'
                 | 'success'
-                | 'needsEmailVerification'
                 | 'needsSmsOtp'
                 | 'needsMfa'
                 | 'failed'
@@ -294,7 +287,6 @@ export interface Typegen0 {
                 | 'passwordlessSms'
                 | 'passwordlessSmsOtp'
                 | 'password'
-                | 'token'
                 | 'anonymous'
                 | 'mfa'
                 | { mfa?: 'totp' }
@@ -312,6 +304,7 @@ export interface Typegen0 {
                   }
             }
         token?: 'idle' | 'running' | { idle?: 'noErrors' | 'error' }
+        email?: 'awaitingVerification' | 'valid'
       }
-  tags: 'ready'
+  tags: 'loading'
 }