+++ packages/hasura-auth-js/src/hasura-auth-client.ts
@@ -119,7 +119,12 @@ export class HasuraAuthClient {
     return new Promise((resolve) => {
       interpreter.send('SIGNUP_EMAIL_PASSWORD', { email, password, options })
       interpreter.onTransition((state) => {
-        if (state.matches({ authentication: { signedOut: 'needsEmailVerification' } })) {
+        if (
+          state.matches({
+            authentication: { signedOut: 'noErrors' },
+            email: 'awaitingVerification'
+          })
+        ) {
           return resolve({ session: null, error: null })
         } else if (state.matches({ authentication: { signedOut: 'failed' } })) {
           return resolve({ session: null, error: state.context.errors.registration || null })
@@ -206,7 +211,12 @@ export class HasuraAuthClient {
               mfa: null,
               error: null
             })
-          } else if (state.matches({ authentication: { signedOut: 'needsEmailVerification' } })) {
+          } else if (
+            state.matches({
+              authentication: { signedOut: 'noErrors' },
+              email: 'awaitingVerification'
+            })
+          ) {
             resolve({
               session: null,
               mfa: null,
@@ -234,7 +244,12 @@ export class HasuraAuthClient {
       return new Promise((resolve) => {
         interpreter.send('SIGNIN_PASSWORDLESS_EMAIL', params)
         interpreter.onTransition((state) => {
-          if (state.matches({ authentication: { signedOut: 'needsEmailVerification' } })) {
+          if (
+            state.matches({
+              authentication: { signedOut: 'noErrors' },
+              email: 'awaitingVerification'
+            })
+          ) {
             resolve({
               session: null,
               mfa: null,
@@ -294,12 +309,25 @@ export class HasuraAuthClient {
         })
       })
     }
-    // TODO anonymous sign-in
-    return {
-      session: null,
-      mfa: null,
-      error: { message: 'Incorrect parameters', status: 500 }
+    // * Anonymous sign-in
+    const { changed } = interpreter.send('SIGNIN_ANONYMOUS')
+    if (!changed) {
+      return { session: null, mfa: null, error: { message: 'already signed in', status: 1 } }
     }
+    return new Promise((resolve) => {
+      interpreter.onTransition((state) => {
+        if (state.matches({ authentication: 'signedIn' })) {
+          resolve({ session: getSession(state.context), mfa: null, error: null })
+        }
+        if (state.matches({ authentication: { signedOut: 'failed' } })) {
+          resolve({
+            session: null,
+            mfa: null,
+            error: state.context.errors.authentication || null
+          })
+        }
+      })
+    })
   }
 
   /**
@@ -745,7 +773,7 @@ export class HasuraAuthClient {
     if (!interpreter) {
       throw Error('Auth interpreter not set')
     }
-    if (interpreter.state.hasTag('ready')) {
+    if (!interpreter.state.hasTag('loading')) {
       return Promise.resolve(interpreter)
     }
     return new Promise((resolve, reject) => {
@@ -754,7 +782,7 @@ export class HasuraAuthClient {
         TIMEOUT_IN_SECONS * 1_000
       )
       interpreter.onTransition((state) => {
-        if (state.hasTag('ready')) {
+        if (!state.hasTag('loading')) {
           clearTimeout(timer)
           return resolve(interpreter)
         }
@@ -763,7 +791,7 @@ export class HasuraAuthClient {
   }
 
   private isReady() {
-    return !!this._client.interpreter?.state?.hasTag('ready')
+    return !this._client.interpreter?.state?.hasTag('loading')
   }
 
   get client() {